Panel: add more warnings
Also:

- Rename the component to UpdateTokenPanel.
- Change the label to be more explicit: “Number of tokens to remove”.
